Add meson and cmake build system to replace qmake
