+++ CMakeLists.txt
@@ -3,10 +3,17 @@ project(Cactus CXX)
 
 set(CMAKE_CXX_STANDARD 14)
 
+# Instruct CMake to run moc automatically when needed
 set(CMAKE_AUTOMOC ON)
 
-find_package(Qt5 COMPONENTS Core Widgets Gui REQUIRED)
+
+
+find_package(Qt5 COMPONENTS Core Widgets Gui 3DCore 3DInput 3DRender 3DExtras REQUIRED)
+
+
 find_package(OpenMVG)
+
+
 #find_package(OpenMVG REQUIRED)
 
 add_subdirectory(src)
