--- ProjetPGAI/imports.qml
@@ -1,10 +1,0 @@
-import QtQuick 2.0
-import QtQuick.Scene3D 2.0
-import Qt3D.Core 2.0
-import Qt3D.Render 2.0
-import Qt3D.Input 2.0
-import Qt3D.Extras 2.0
-
-Item {
-
-}
