+++ src/gui/mainwindow.h
@@ -1,11 +1,13 @@
 #ifndef MAINWINDOW_H
 #define MAINWINDOW_H
 
-#include <QMainWindow>
-#include <QtWidgets>
 #include "objwindow.h"
 
-namespace Ui {
+#include <QMainWindow>
+#include <QtWidgets>
+
+namespace Ui
+{
 class MainWindow;
 }
 
@@ -14,7 +16,7 @@ class MainWindow : public QMainWindow
     Q_OBJECT
 
 public:
-    explicit MainWindow(QWidget *parent = nullptr);
+    explicit MainWindow( QWidget *parent = nullptr );
     QPushButton *btn1;
     ~MainWindow();
 
@@ -23,7 +25,7 @@ private slots:
 
 private:
     Ui::MainWindow *ui;
-    OBJWindow * owin;
+    OBJWindow *owin;
 };
 
 #endif // MAINWINDOW_H
