+++ src/gui/objwindow.cpp
